Simplify hot-list URL and set default page in 36kr
In radar.js, the '/hot-list/:category/:date/:page' source URL was simplified to '/hot-list/:category' to decrease the need for argument data that is often irrelevant or constant. In hot-list.js, a default page of 1 was set and inserted into the URL's path to accommodate this change. This makes the URLs easier to handle and reduces the potential for errors.